+++ test/__tests__/twitter_ads.test.js
@@ -13,6 +13,20 @@ const testDataFile = fs.readFileSync(
 );
 const testData = JSON.parse(testDataFile);
 
+const authHeaderConstant = "OAuth oauth_consumer_key=\"qwe\", oauth_nonce=\"V1kMh028kZLLhfeYozuL0B45Pcx6LvuW\", oauth_signature=\"Di4cuoGv4PnCMMEeqfWTcqhvdwc%3D\", oauth_signature_method=\"HMAC-SHA1\", oauth_timestamp=\"1685603652\", oauth_token=\"yrdghfvhjvhj\", oauth_version=\"1.0\"";
+
+jest.mock("../../src/v0/destinations/twitter_ads/util", () => {
+  const originalModule = jest.requireActual("../../src/v0/destinations/twitter_ads/util");
+  return {
+    ...originalModule,
+    getAuthHeaderForRequest: jest.fn(() => {
+      return {
+        Authorization: authHeaderConstant
+      }
+    })
+  }
+})
+
 describe(`${name} Tests`, () => {
   describe("Processor", () => {
     testData.forEach((dataPoint, index) => {
